move getTopicByTag in simplifionstopicsApi
