+++ src/custom/simplifions/components/SimplifionsRecoSolutions.vue
@@ -191,7 +191,6 @@
 </template>
 
 <script setup lang="ts">
-import TopicsAPI from '@/services/api/resources/TopicsAPI.ts'
 import { fromMarkdown } from '@/utils'
 import { grist } from '../grist.ts'
 import type {
@@ -200,6 +199,7 @@ import type {
   Recommandation,
   SolutionRecord
 } from '../model/grist'
+import TopicsAPI from '../simplifionsTopicsApi'
 
 const props = defineProps<{
   recommandation: Recommandation
